feat: Build a docker image for Boon.
Boon is https://github.com/santhosh-tekuri/boon

It is a tool to validate json schema and to verify json files against
this schema. Alternatives are ajv and https://python-jsonschema.readthedocs.io
ajv needs typescript and nodejs. And when installed, is picky about the
schema: the schema we "Get" from imsglobal, it defines as invalid, for
example. python-jsonschema requires a particular python runtime with ssl
setup correct. Difficult and fragile.

Hence the simple rust tool "boon".

We build it into a docker container and push that to our own registry so
that we can use it in the CI