feat: Define contexts according to OpenBadges 3.0 Implementation Guide
https://www.imsglobal.org/spec/ob/v3p0/impl/#open-badges-extensions
This explains that we cannot just include contexts. Instead, we must
define our own type of Achievement and include them there.

We therefore introduce an ExtracurricularAchievement type. This extends
the Achievement by adding the extended fields.

For now, we've implemented the additional fields in this one file. In
future we may want to share field definitions between other types and we
can do this with impprted-contexts in JSON-ld 1.1 https://www.w3.org/TR/json-ld11/#imported-contexts
But that
 - introduces a requirement for version 1.1
 - adds a lot of noise.
 - requires more json to define all the linking and importing than we
   have now with hardcoding them.