feat: Move to an Authorization Code Flow issuer
This allows us to move away from the "poc9" issuer. And to test
with an Authorization Code Flow. Most wallets, by now *only* support
this, so having that makes sense.

We'll move back to an impierce ssi-agent issuer once we have stable one
running on a public SDP tier (playground?) and once that supports the
authorization code flow fully.